+++ app/src/main/java/fr/scootop/app/services/ApiConfigurationService.kt
@@ -1,8 +1,6 @@
 package fr.scootop.app.services
 
 import fr.scootop.data.model.configuration.*
-
-import retrofit2.http.Headers
 import retrofit2.Response
 import retrofit2.http.*
 
@@ -16,7 +14,7 @@ interface ApiConfigurationService {
 
     @Headers("Content-Type: application/json")
     @GET("/configuration/action/type/{id}")
-    fun getActionTypeById(@Path("id") id: Integer): Response<ActionType>
+    fun getActionTypeById(@Path("id") id: Long): Response<ActionType>
 
     @Headers("Content-Type: application/json")
     @POST("/configuration/action/type")
@@ -24,11 +22,11 @@ interface ApiConfigurationService {
 
     @Headers("Content-Type: application/json")
     @PUT("/configuration/action/type/{id}")
-    fun updateActionTypeById(@Path("id") id: Integer, @Body actionType: ActionType): Response<ActionType>
+    fun updateActionTypeById(@Path("id") id: Long, @Body actionType: ActionType): Response<ActionType>
 
     @Headers("Content-Type: application/json")
     @DELETE("/configuration/action/type/{id}")
-    fun deleteActionType(@Path("id") id: Integer): Response<String>
+    fun deleteActionType(@Path("id") id: Long): Response<String>
 
     // CATEGORY
 
@@ -38,7 +36,7 @@ interface ApiConfigurationService {
 
     @Headers("Content-Type: application/json")
     @GET("/configuration/category/{id}")
-    fun getCategoryById(@Path("id") id: Integer): Response<Category>
+    fun getCategoryById(@Path("id") id: Long): Response<Category>
 
     @Headers("Content-Type: application/json")
     @POST("/configuration/category")
@@ -46,17 +44,17 @@ interface ApiConfigurationService {
 
     @Headers("Content-Type: application/json")
     @PUT("/configuration/category/{id}")
-    fun updateCategory(@Path("id") id: Integer, @Body category: Category): Response<Category>
+    fun updateCategory(@Path("id") id: Long, @Body category: Category): Response<Category>
 
     @Headers("Content-Type: application/json")
     @DELETE("/configuration/category/{id}")
-    fun deleteCategory(@Path("id") id: Integer): Response<String>
+    fun deleteCategory(@Path("id") id: Long): Response<String>
 
     //CATEGORY TYPE
 
     @Headers("Content-Type: application/json")
     @GET("/configuration/category/type/{id}")
-    fun getCategoryTypeById(@Path("id") id: Integer): Response<CategoryType>
+    fun getCategoryTypeById(@Path("id") id: Long): Response<CategoryType>
 
     @Headers("Content-Type: application/json")
     @GET("/configuration/category/types")
@@ -68,18 +66,18 @@ interface ApiConfigurationService {
 
     @Headers("Content-Type: application/json")
     @PUT("/configuration/category/type/{id}")
-    fun updateCategoryType(@Path("id") id: Integer, @Body categoryType: CategoryType): Response<CategoryType>
+    fun updateCategoryType(@Path("id") id: Long, @Body categoryType: CategoryType): Response<CategoryType>
 
     @Headers("Content-Type: application/json")
     @DELETE("/configuration/category/type/{id}")
-    fun deleteCategoryType(@Path("id") id: Integer): Response<String>
+    fun deleteCategoryType(@Path("id") id: Long): Response<String>
 
 
     //COMPETITION TYPE
 
     @Headers("Content-Type: application/json")
     @GET("/configuration/competition/type/{id}")
-    fun getCompetitionTypeById(@Path("id") id: Integer): Response<CompetitionType>
+    fun getCompetitionTypeById(@Path("id") id: Long): Response<CompetitionType>
 
     @Headers("Content-Type: application/json")
     @GET("/configuration/competition/types")
@@ -91,17 +89,17 @@ interface ApiConfigurationService {
 
     @Headers("Content-Type: application/json")
     @PUT("/configuration/competition/type/{id}")
-    fun updateCompetitionType(@Path("id") id: Integer, @Body competitionType: CompetitionType): Response<CompetitionType>
+    fun updateCompetitionType(@Path("id") id: Long, @Body competitionType: CompetitionType): Response<CompetitionType>
 
     @Headers("Content-Type: application/json")
     @DELETE("/configuration/competition/type/{id}")
-    fun deleteCompetitionType(@Path("id") id: Integer): Response<String>
+    fun deleteCompetitionType(@Path("id") id: Long): Response<String>
 
     // DIVISION
 
     @Headers("Content-Type: application/json")
     @GET("/configuration/division/{id}")
-    fun getDivisionById(@Path("id") id: Integer): Response<Division>
+    fun getDivisionById(@Path("id") id: Long): Response<Division>
 
     @Headers("Content-Type: application/json")
     @GET("/configuration/divisions")
@@ -113,17 +111,17 @@ interface ApiConfigurationService {
 
     @Headers("Content-Type: application/json")
     @PUT("/configuration/division/{id}")
-    fun updateDivision(@Path("id") id: Integer, @Body division: Division): Response<Division>
+    fun updateDivision(@Path("id") id: Long, @Body division: Division): Response<Division>
 
     @Headers("Content-Type: application/json")
     @DELETE("/configuration/division/{id}")
-    fun deleteDivision(@Path("id") id: Integer): Response<String>
+    fun deleteDivision(@Path("id") id: Long): Response<String>
 
     // LEAGUE
 
     @Headers("Content-Type: application/json")
     @GET("/configuration/league/{id}")
-    fun getLeagueById(@Path("id") id: Integer): Response<League>
+    fun getLeagueById(@Path("id") id: Long): Response<League>
 
     @Headers("Content-Type: application/json")
     @GET("/configuration/leagues")
@@ -135,17 +133,17 @@ interface ApiConfigurationService {
 
     @Headers("Content-Type: application/json")
     @PUT("/configuration/league/{id}")
-    fun updateLeague(@Path("id") id: Integer, @Body league: League): Response<League>
+    fun updateLeague(@Path("id") id: Long, @Body league: League): Response<League>
 
     @Headers("Content-Type: application/json")
     @DELETE("/configuration/league/{id}")
-    fun deleteLeague(@Path("id") id: Integer): Response<String>
+    fun deleteLeague(@Path("id") id: Long): Response<String>
 
     // POSTE
 
     @Headers("Content-Type: application/json")
     @GET("/configuration/poste/{id}")
-    fun getPosteById(@Path("id") id: Integer): Response<Poste>
+    fun getPosteById(@Path("id") id: Long): Response<Poste>
 
     @Headers("Content-Type: application/json")
     @GET("/configuration/postes")
@@ -157,17 +155,17 @@ interface ApiConfigurationService {
 
     @Headers("Content-Type: application/json")
     @PUT("/configuration/poste/{id}")
-    fun updatePoste(@Path("id") id: Integer, @Body poste: Poste): Response<Poste>
+    fun updatePoste(@Path("id") id: Long, @Body poste: Poste): Response<Poste>
 
     @Headers("Content-Type: application/json")
     @DELETE("/configuration/poste/{id}")
-    fun deletePoste(@Path("id") id: Integer): Response<String>
+    fun deletePoste(@Path("id") id: Long): Response<String>
 
     // STADE
 
     @Headers("Content-Type: application/json")
     @GET("/configuration/stade/{id}")
-    fun getStadeById(@Path("id") id: Integer): Response<Stade>
+    fun getStadeById(@Path("id") id: Long): Response<Stade>
 
     @Headers("Content-Type: application/json")
     @GET("/configuration/stades")
@@ -179,17 +177,17 @@ interface ApiConfigurationService {
 
     @Headers("Content-Type: application/json")
     @PUT("/configuration/stade/{id}")
-    fun updateStade(@Path("id") id: Integer, @Body stade: Stade): Response<Stade>
+    fun updateStade(@Path("id") id: Long, @Body stade: Stade): Response<Stade>
 
     @Headers("Content-Type: application/json")
     @DELETE("/configuration/stade/{id}")
-    fun deleteStade(@Path("id") id: Integer): Response<String>
+    fun deleteStade(@Path("id") id: Long): Response<String>
 
     //TOUCH TYPE
 
     @Headers("Content-Type: application/json")
     @GET("/configuration/touch/type/{id}")
-    fun getTouchTypeById(@Path("id") id: Integer): Response<TouchType>
+    fun getTouchTypeById(@Path("id") id: Long): Response<TouchType>
 
     @Headers("Content-Type: application/json")
     @GET("/configuration/touch/types")
@@ -201,10 +199,10 @@ interface ApiConfigurationService {
 
     @Headers("Content-Type: application/json")
     @PUT("/configuration/touch/type/{id}")
-    fun updateTouchType(@Path("id") id: Integer, @Body touchType: TouchType): Response<TouchType>
+    fun updateTouchType(@Path("id") id: Long, @Body touchType: TouchType): Response<TouchType>
 
     @Headers("Content-Type: application/json")
     @DELETE("/configuration/touch/type/{id}")
-    fun deleteTouchType(@Path("id") id: Integer): Response<String>
+    fun deleteTouchType(@Path("id") id: Long): Response<String>
 
 }
