+++ app/src/main/java/fr/scootop/app/services/ApiToolsService.kt
@@ -1,7 +1,6 @@
 package fr.scootop.app.services
 
 import fr.scootop.data.model.tools.*
-import retrofit2.http.Headers
 import retrofit2.Response
 import retrofit2.http.*
 
@@ -15,7 +14,7 @@ interface ApiToolsService {
 
     @Headers("Content-Type: application/json")
     @GET("/tools/club/{id}")
-    fun getClubById(@Path("id") id: Integer): Response<Club>
+    fun getClubById(@Path("id") id: Long): Response<Club>
 
     @Headers("Content-Type: application/json")
     @POST("/tools/club")
@@ -23,11 +22,11 @@ interface ApiToolsService {
 
     @Headers("Content-Type: application/json")
     @PUT("/tools/club/{id}")
-    fun updateClub(@Path("id") id: Integer, @Body club: Club): Response<Club>
+    fun updateClub(@Path("id") id: Long, @Body club: Club): Response<Club>
 
     @Headers("Content-Type: application/json")
     @DELETE("/tools/club/{id}")
-    fun deleteClub(@Path("id") id: Integer): Response<String>
+    fun deleteClub(@Path("id") id: Long): Response<String>
 
     // COMPETITION
 
@@ -37,7 +36,7 @@ interface ApiToolsService {
 
     @Headers("Content-Type: application/json")
     @GET("/tools/competition/{id}")
-    fun getCompetitionById(@Path("id") id: Integer): Response<Competition>
+    fun getCompetitionById(@Path("id") id: Long): Response<Competition>
 
     @Headers("Content-Type: application/json")
     @POST("/tools/competition")
@@ -45,11 +44,11 @@ interface ApiToolsService {
 
     @Headers("Content-Type: application/json")
     @PUT("/tools/competition/{id}")
-    fun updateCompetition(@Path("id") id: Integer, @Body competition: Competition): Response<Competition>
+    fun updateCompetition(@Path("id") id: Long, @Body competition: Competition): Response<Competition>
 
     @Headers("Content-Type: application/json")
     @DELETE("/tools/competition/{id}")
-    fun deleteCompetition(@Path("id") id: Integer): Response<String>
+    fun deleteCompetition(@Path("id") id: Long): Response<String>
 
     // MATCH
 
@@ -59,7 +58,7 @@ interface ApiToolsService {
 
     @Headers("Content-Type: application/json")
     @GET("/tools/match/{id}")
-    fun getMatchById(@Path("id") id: Integer): Response<Match>
+    fun getMatchById(@Path("id") id: Long): Response<Match>
 
     @Headers("Content-Type: application/json")
     @POST("/tools/match")
@@ -67,11 +66,11 @@ interface ApiToolsService {
 
     @Headers("Content-Type: application/json")
     @PUT("/tools/match/{id}")
-    fun updateMatch(@Path("id") id: Integer, @Body match: Match): Response<Match>
+    fun updateMatch(@Path("id") id: Long, @Body match: Match): Response<Match>
 
     @Headers("Content-Type: application/json")
     @DELETE("/tools/match/{id}")
-    fun deleteMatch(@Path("id") id: Integer): Response<String>
+    fun deleteMatch(@Path("id") id: Long): Response<String>
 
 
     // SAISON
@@ -82,7 +81,7 @@ interface ApiToolsService {
 
     @Headers("Content-Type: application/json")
     @GET("/tools/saison/{id}")
-    fun getSaisonById(@Path("id") id: Integer): Response<Saison>
+    fun getSaisonById(@Path("id") id: Long): Response<Saison>
 
     @Headers("Content-Type: application/json")
     @POST("/tools/saison")
@@ -90,11 +89,11 @@ interface ApiToolsService {
 
     @Headers("Content-Type: application/json")
     @PUT("/tools/saison/{id}")
-    fun updateSaison(@Path("id") id: Integer, @Body saison: Saison): Response<Saison>
+    fun updateSaison(@Path("id") id: Long, @Body saison: Saison): Response<Saison>
 
     @Headers("Content-Type: application/json")
     @DELETE("/tools/saison/{id}")
-    fun deleteSaison(@Path("id") id: Integer): Response<String>
+    fun deleteSaison(@Path("id") id: Long): Response<String>
 
 
     // STAT
@@ -105,7 +104,7 @@ interface ApiToolsService {
 
     @Headers("Content-Type: application/json")
     @GET("/tools/statisticalsheet/{id}")
-    fun getStatisticalSheetById(@Path("id") id: Integer): Response<StatisticalSheet>
+    fun getStatisticalSheetById(@Path("id") id: Long): Response<StatisticalSheet>
 
     @Headers("Content-Type: application/json")
     @POST("/tools/statisticalsheet")
@@ -113,15 +112,15 @@ interface ApiToolsService {
 
     @Headers("Content-Type: application/json")
     @PUT("/tools/statisticalsheet/{id}")
-    fun updateStatisticalSheett(@Path("id") id: Integer, @Body statisticalSheet: StatisticalSheet): Response<StatisticalSheet>
+    fun updateStatisticalSheett(@Path("id") id: Long, @Body statisticalSheet: StatisticalSheet): Response<StatisticalSheet>
 
     @Headers("Content-Type: application/json")
     @DELETE("/tools/statisticalsheet/{id}")
-    fun deleteStatisticalSheet(@Path("id") id: Integer): Response<String>
+    fun deleteStatisticalSheet(@Path("id") id: Long): Response<String>
 
     @Headers("Content-Type: application/json")
     @GET("/tools/stats/{playerId}/{eventId}")
-    fun getStatByPLayerAndEvent(@Path("playerId") playerId: Integer, @Path("eventId") eventId: Integer): Response<StatisticalSheet>
+    fun getStatByPLayerAndEvent(@Path("playerId") playerId: Long, @Path("eventId") eventId: Long): Response<StatisticalSheet>
 
 
 
@@ -133,7 +132,7 @@ interface ApiToolsService {
 
     @Headers("Content-Type: application/json")
     @GET("/tools/team/{id}")
-    fun getTeamById(@Path("id") id: Integer): Response<Team>
+    fun getTeamById(@Path("id") id: Long): Response<Team>
 
     @Headers("Content-Type: application/json")
     @POST("/tools/team")
@@ -141,10 +140,10 @@ interface ApiToolsService {
 
     @Headers("Content-Type: application/json")
     @PUT("/tools/team/{id}")
-    fun updateTeams(@Path("id") id: Integer, @Body team: Team): Response<Team>
+    fun updateTeams(@Path("id") id: Long, @Body team: Team): Response<Team>
 
     @Headers("Content-Type: application/json")
     @DELETE("/tools/tam/{id}")
-    fun deleteTeam(@Path("id") id: Integer): Response<String>
+    fun deleteTeam(@Path("id") id: Long): Response<String>
 
 }
